+++ ogma-core/src/Command/Standalone.hs
@@ -42,49 +42,24 @@ module Command.Standalone
   where
 
 -- External imports
-import Control.Exception      as E
-import Control.Monad.Except   (ExceptT (..), liftEither, runExceptT, throwError)
-import Control.Monad.IO.Class (liftIO)
-import Data.Aeson             (ToJSON (..), Value (Null, Object), decode,
-                               eitherDecode, object)
-import Data.Aeson.KeyMap      (union)
-import Data.ByteString.Lazy   (fromStrict)
-import Data.List              (isInfixOf, isPrefixOf, nub, (\\))
-import Data.Maybe             (fromMaybe)
-import GHC.Generics           (Generic)
-import System.Directory       (doesFileExist)
-import System.FilePath        ((</>))
-import System.Process         (readProcess)
+import Control.Exception    as E
+import Control.Monad.Except (ExceptT (..), liftEither)
+import Data.Aeson           (ToJSON (..))
+import Data.List            (nub, (\\))
+import Data.Maybe           (fromMaybe)
+import GHC.Generics         (Generic)
 
--- External imports: auxiliary
-import Data.ByteString.Extra  as B ( safeReadFile )
-import System.Directory.Extra ( copyTemplate )
+-- External imports: Ogma
+import Data.OgmaSpec          (ExternalVariableDef (..),
+                               InternalVariableDef (..), Requirement (..),
+                               Spec (..))
+import System.Directory.Extra (copyTemplate)
 
--- External imports: language ASTs, transformers
-import Data.OgmaSpec (ExternalVariableDef (..), InternalVariableDef (..),
-                      Requirement (..), Spec (..))
-import Language.JSONSpec.Parser (JSONFormat (..), parseJSONSpec)
-import Language.XMLSpec.Parser  (parseXMLSpec)
-
--- External imports: language ASTs, transformers
-import qualified Language.CoCoSpec.AbsCoCoSpec as CoCoSpec
-import qualified Language.CoCoSpec.ParCoCoSpec as CoCoSpec ( myLexer,
-                                                             pBoolSpec )
-
-import qualified Language.SMV.AbsSMV       as SMV
-import qualified Language.SMV.ParSMV       as SMV (myLexer, pBoolSpec)
-import           Language.SMV.Substitution (substituteBoolExpr)
-
-import qualified Language.Trans.CoCoSpec2Copilot as CoCoSpec (boolSpec2Copilot,
-                                                              boolSpecNames)
-import           Language.Trans.SMV2Copilot      as SMV (boolSpec2Copilot,
-                                                         boolSpecNames)
-import           Language.Trans.Spec2Copilot     (spec2Copilot, specAnalyze)
-
--- Internal imports: auxiliary
-import Command.Result  (Result (..))
-import Data.Location   (Location (..))
-import Paths_ogma_core (getDataDir)
+-- Internal imports
+import Command.Common
+import Command.Result              (Result (..))
+import Data.Location               (Location (..))
+import Language.Trans.Spec2Copilot (spec2Copilot, specAnalyze)
 
 -- | Generate a new standalone Copilot monitor that implements the spec in an
 -- input file.
@@ -210,215 +185,6 @@ data AppData = AppData
 
 instance ToJSON AppData
 
--- | Process input specification, if available, and return its abstract
--- representation.
-parseInputFile :: FilePath
-               -> String
-               -> String
-               -> Maybe String
-               -> ExprPairT a
-               -> ExceptT ErrorTriplet IO (Spec a)
-parseInputFile fp formatName propFormatName propVia exprT =
-  ExceptT $ do
-    let ExprPairT parse replace print ids def = exprT
-
-    let wrapper = wrapVia propVia parse
-    -- Obtain format file.
-    --
-    -- A format name that exists as a file in the disk always takes preference
-    -- over a file format included with Ogma. A file format with a forward
-    -- slash in the name is always assumed to be a user-provided filename.
-    -- Regardless of whether the file is user-provided or known to Ogma, we
-    -- check (again) whether the file exists, and print an error message if
-    -- not.
-    exists  <- doesFileExist formatName
-    dataDir <- getDataDir
-    let formatFile
-          | isInfixOf "/" formatName || exists
-          = formatName
-          | otherwise
-          = dataDir </> "data" </> "formats" </>
-               (formatName ++ "_" ++ propFormatName)
-    formatMissing <- not <$> doesFileExist formatFile
-
-    if formatMissing
-      then return $ Left $ commandIncorrectFormatSpec formatFile
-      else do
-        res <- do
-          format <- readFile formatFile
-
-          -- All of the following operations use Either to return error
-          -- messages.  The use of the monadic bind to pass arguments from one
-          -- function to the next will cause the program to stop at the
-          -- earliest error.
-          if | isPrefixOf "XMLFormat" format
-             -> do let xmlFormat = read format
-                   content <- readFile fp
-                   parseXMLSpec
-                     (wrapper) (def) xmlFormat content
-                     -- (fmap (fmap print) . wrapper) (print def) xmlFormat content
-             | otherwise
-             -> do let jsonFormat = read format
-                   content <- B.safeReadFile fp
-                   case content of
-                     Left e  -> return $ Left e
-                     Right b -> do case eitherDecode b of
-                                     Left e  -> return $ Left e
-                                     Right v ->
-                                       parseJSONSpec
-                                         (wrapper)
-                                         jsonFormat
-                                         v
-        case res of
-          Left _  -> return $ Left $ cannotOpenInputFile fp
-          Right x -> return $ Right x
-
--- | Process a JSON file with additional template variables to make available
--- during template expansion.
-parseTemplateVarsFile :: Maybe FilePath
-                      -> ExceptT ErrorTriplet IO Value
-parseTemplateVarsFile Nothing   = return $ object []
-parseTemplateVarsFile (Just fp) = do
-  content <- liftIO $ B.safeReadFile fp
-  let value = eitherDecode =<< content
-  case value of
-    Right x@(Object _) -> return x
-    Right x@Null       -> return x
-    Right _            -> throwError (cannotReadObjectTemplateVars fp)
-    _                  -> throwError (cannotOpenTemplateVars fp)
-
--- * Handler for boolean expressions
-
--- | Handler for boolean expressions that knows how to parse them, replace
--- variables in them, and convert them to Copilot.
---
--- It also contains a default value to be used whenever an expression cannot be
--- found in the input file.
-data ExprPair = forall a . ExprPair
-  { exprTPair   :: ExprPairT a
-  }
-
-data ExprPairT a = ExprPairT
-  { exprTParse   :: String -> Either String a
-  , exprTReplace :: [(String, String)] -> a -> a
-  , exprTPrint   :: a -> String
-  , exprTIdents  :: a -> [String]
-  , exprTUnknown :: a
-  }
-
-
--- | Return a handler depending on whether it should be for CoCoSpec boolean
--- expressions or for SMV boolean expressions. We default to SMV if not format
--- is given.
-exprPair :: String -> ExprPair
-exprPair "cocospec" = ExprPair $
-  ExprPairT
-    (CoCoSpec.pBoolSpec . CoCoSpec.myLexer)
-    (\_ -> id)
-    (CoCoSpec.boolSpec2Copilot)
-    (CoCoSpec.boolSpecNames)
-    (CoCoSpec.BoolSpecSignal (CoCoSpec.Ident "undefined"))
-exprPair "literal" = ExprPair $
-  ExprPairT
-    Right
-    (\_ -> id)
-    id
-    (const [])
-    "undefined"
-exprPair _ = ExprPair $
-  ExprPairT
-    (SMV.pBoolSpec . SMV.myLexer)
-    (substituteBoolExpr)
-    (SMV.boolSpec2Copilot)
-    (SMV.boolSpecNames)
-    (SMV.BoolSpecSignal (SMV.Ident "undefined"))
-
--- | Parse a property using an auxiliary program to first translate it, if
--- available.
---
--- If a program is given, it is first called on the property, and then the
--- result is parsed with the parser passed as an argument. If a program is not
--- given, then the parser is applied to the given string.
-wrapVia :: Maybe String                -- ^ Auxiliary program to translate the
-                                       -- property.
-        -> (String -> Either String a) -- ^ Parser used on the result.
-        -> String                      -- ^ Property to parse.
-        -> IO (Either String a)
-wrapVia Nothing  parse s = return (parse s)
-wrapVia (Just f) parse s =
-  E.handle (\(e :: E.IOException) -> return $ Left $ show e) $ do
-    out <- readProcess f [] s
-    return $ parse out
-
--- * Errors
-
--- | A triplet containing error information.
-data ErrorTriplet = ErrorTriplet ErrorCode String Location
-
--- | Encoding of reasons why the command can fail.
---
--- The error codes used are 1 for user error, and 2 for internal bug.
-type ErrorCode = Int
-
--- | Process a computation that can fail with an error code, and turn it into a
--- computation that returns a 'Result'.
-processResult :: Monad m => ExceptT ErrorTriplet m a -> m (Result ErrorCode)
-processResult m = do
-  r <- runExceptT m
-  case r of
-    Left (ErrorTriplet errorCode msg location)
-      -> return $ Error errorCode msg location
-    _ -> return Success
-
--- ** Error messages
--- | Exception handler to deal with the case in which the input file cannot be
--- opened.
-cannotOpenInputFile :: FilePath -> ErrorTriplet
-cannotOpenInputFile file =
-    ErrorTriplet ecCannotOpenInputFile msg (LocationFile file)
-  where
-    msg =
-      "cannot open input specification file " ++ file
-
-
--- | Error message associated to the format file not being found.
-commandIncorrectFormatSpec :: FilePath -> ErrorTriplet
-commandIncorrectFormatSpec formatFile =
-    ErrorTriplet ecIncorrectFormatFile msg (LocationFile formatFile)
-  where
-    msg =
-      "The format specification " ++ formatFile ++ " does not exist or is not "
-      ++ "readable"
-
--- | Exception handler to deal with the case in which the template vars file
--- cannot be opened.
-cannotOpenTemplateVars :: FilePath -> ErrorTriplet
-cannotOpenTemplateVars file =
-    ErrorTriplet ecCannotOpenTemplateVarsFile msg (LocationFile file)
-  where
-    msg =
-      "Cannot open file with additional template variables: " ++ file
-
--- | Exception handler to deal with the case in which the template vars file
--- cannot be opened.
-cannotReadObjectTemplateVars :: FilePath -> ErrorTriplet
-cannotReadObjectTemplateVars file =
-    ErrorTriplet ecCannotReadObjectTemplateVarsFile msg (LocationFile file)
-  where
-    msg =
-      "Cannot open file with additional template variables: " ++ file
-
--- | Exception handler to deal with the case of files that cannot be
--- copied/generated due lack of space or permissions or some I/O error.
-cannotCopyTemplate :: ErrorTriplet
-cannotCopyTemplate =
-    ErrorTriplet ecCannotCopyTemplate msg LocationNothing
-  where
-    msg =
-      "Generation failed during copy/write operation. Check that"
-      ++ " there's free space in the disk and that you have the necessary"
-      ++ " permissions to write in the destination directory."
-
 -- | Error message associated to not being able to formalize the input spec.
 commandIncorrectSpec :: String -> String -> ErrorTriplet
 commandIncorrectSpec file e =
@@ -429,30 +195,9 @@ commandIncorrectSpec file e =
 
 -- ** Error codes
 
--- | Error: the input specification provided by the user cannot be opened.
-ecCannotOpenInputFile :: ErrorCode
-ecCannotOpenInputFile = 1
-
 -- | Error: the input specification cannot be formalized.
 ecIncorrectSpec :: ErrorCode
 ecIncorrectSpec = 1
-
--- | Error: the format file cannot be opened.
-ecIncorrectFormatFile :: ErrorCode
-ecIncorrectFormatFile = 1
-
--- | Error: the template vars file provided by the user cannot be opened.
-ecCannotOpenTemplateVarsFile :: ErrorCode
-ecCannotOpenTemplateVarsFile = 1
-
--- | Error: the template variables file passed does not contain a JSON object.
-ecCannotReadObjectTemplateVarsFile :: ErrorCode
-ecCannotReadObjectTemplateVarsFile = 1
-
--- | Error: the files cannot be copied/generated due lack of space or
--- permissions or some I/O error.
-ecCannotCopyTemplate :: ErrorCode
-ecCannotCopyTemplate = 1
 
 -- | Add to a spec external variables for all identifiers mentioned in
 -- expressions that are not defined anywhere.
@@ -471,31 +216,6 @@ addMissingIdentifiers f s = s { externalVariables = vars' }
     -- Names that are defined in variables.
     existingNames = map externalVariableName (externalVariables s)
                  ++ map internalVariableName (internalVariables s)
-
--- | Return the path to the template directory.
-locateTemplateDir :: Maybe FilePath
-                  -> FilePath
-                  -> ExceptT e IO FilePath
-locateTemplateDir mTemplateDir name =
-  case mTemplateDir of
-    Just x  -> return x
-    Nothing -> liftIO $ do
-      dataDir <- getDataDir
-      return $ dataDir </> "templates" </> name
-
--- | Merge two JSON objects.
---
--- Fails if the values are not objects or null.
-mergeObjects :: Value -> Value -> Value
-mergeObjects (Object m1) (Object m2) = Object (union m1 m2)
-mergeObjects obj         Null        = obj
-mergeObjects Null        obj         = obj
-mergeObjects _           _           = error "The values passed are not objects"
-
--- | Replace the left Exception in an Either.
-makeLeftE :: c -> Either E.SomeException b -> Either c b
-makeLeftE c (Left _)   = Left c
-makeLeftE _ (Right x)  = Right x
 
 mapLeft :: (a -> c) -> Either a b -> Either c b
 mapLeft f (Left x)  = Left (f x)
